Features: Changed name, Removed data processing, Added Regularization, Added output dependent models
Changed name to include the regularization parameter. This makes it
easier to identify individual networks in a list when optimizing.

Removed data processing because regularization was added.

Added Regularization to initialization of the class to follow the
format of previous learner models.

Added output dependent models to automatically build models if the
output is returns rather than prices, and vice versa.