Feature: Added functions for building input and output datasets
Added functions for building input and output datasets. This includes
one function to retrieve data from the yahoo finance API. Another to
populate a data folder with all companies currently in the S&P 500.
Finally, a function to build the feature variable set and a function to
build the output variable.